+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.0.86 (Nov 07, 2022)
+* Added a new `profile` command to output the current profile configuration for the CLI. Use this to help debug any unexpected results from CLI commands.
+* Updated the `envs new` command to accept a new `--preview` parameter. If `--preview` is passed, the default provider and region configured for the stack will be used to configure the new preview environment.
+* Added a new command `envs up` that is used to launch and deploy an entire environment.
+* Added a new command `envs down` that is used to destroy an entire environment.
+* Added a new command `envs delete` that is used to delete an environment once the infrastructure has been destroyed.
+
 # 0.0.85 (Oct 21, 2022)
 * Updated to account for changes in environment ordering and preview environments.
 * Improved logging for runs (`up` command) and deploys (`launch` and `deploy` commands).